requirements.txt
alembic==0.8.7
blinker==1.4
cffi==1.9.1
click==6.6
cryptography==1.7.1
decorator==4.0.10
Flask==0.11.1
Flask-HTTPAuth==3.2.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==2.0.0
Flask-Script==2.0.5
Flask-Session==0.3.0
Flask-SQLAlchemy==2.1
Flask-Webpack==0.1.0
Flask-WTF==0.12
Frozen-Flask==0.13
idna==2.1
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
numpy==1.11.1
pandas==0.18.1
pbr==1.10.0
pyasn1==0.1.9
pycparser==2.17
pycrypto==2.6.1
PyJWT==1.4.2
PyMySQL==0.7.9
python-dateutil==2.5.3
python-editor==1.0.1
pytz==2016.6.1
requests==2.13.0
six==1.10.0
SQLAlchemy==1.0.14
sqlalchemy-migrate==0.10.0
sqlparse==0.2.1
Tempita==0.5.2
Werkzeug==0.11.10
WTForms==2.1
WTForms-ParsleyJS==2.0.1
